+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next version
 
+- add fancy argument parsing
 - add login and logout command to room
 - add pm command to room
 - add cookie support